Fix UI lint so Fast Refresh and hook deps pass CI.
Split sidebar helpers from JSX components and stop reasoning/history effects from depending on values they must not re-run on.
