+++ packages/ui/src/components/session/sidebar/SidebarProjectsList.tsx
@@ -2,7 +2,8 @@ import React from 'react';
 import { ScrollableOverlay } from '@/components/ui/ScrollableOverlay';
 import { formatDirectoryName, cn } from '@/lib/utils';
 import type { SessionGroup, SessionNode } from './types';
-import { formatProjectLabel, SidebarSessionLikeButton } from './utils';
+import { formatProjectLabel } from './utils';
+import { SidebarSessionLikeButton } from './sidebarRowChrome';
 import type { ProjectSortOrder } from '@/stores/useSessionDisplayStore';
 import { streamPerfCount } from '@/stores/utils/streamDebug';
 import type { SortableDragHandleProps } from './sortableItems';
